Move activity to androidMain
