+++ decompose-router/src/androidMain/kotlin/io/github/xxfast/decompose/TestActivity.kt
@@ -1,0 +1,27 @@
+package io.github.xxfast.decompose
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.core.view.WindowCompat
+import com.arkivanov.decompose.DefaultComponentContext
+import com.arkivanov.decompose.defaultComponentContext
+import io.github.xxfast.decompose.screen.HomeScreen
+
+class TestActivity : ComponentActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    WindowCompat.setDecorFitsSystemWindows(window, false)
+    val rootComponentContext: DefaultComponentContext = defaultComponentContext()
+
+    setContent {
+      CompositionLocalProvider(LocalComponentContext provides rootComponentContext) {
+        MaterialTheme {
+          HomeScreen()
+        }
+      }
+    }
+  }
+}
